+++ lib/dark_worlds_server/config/characters/projectile.ex
@@ -13,10 +13,10 @@ defmodule DarkWorldsServer.Config.Characters.Projectile do
   schema "projectiles" do
     field(:name, :string)
     field(:base_damage, :integer)
-    field(:base_speed, :integer)
-    field(:base_size, :integer)
+    field(:base_speed, :float)
+    field(:base_size, :float)
     field(:duration_ms, :integer)
-    field(:max_distance, :integer)
+    field(:max_distance, :float)
     field(:remove_on_collision, :boolean)
 
     many_to_many(:on_hit_effects, Effect, join_through: ProjectileEffect)
